Generate webpack_<entry>.liquid per entry
This is a breaking change:

There is no longer a `webpack.liquid` generated. The default entry is
called `app`, and a corresponding `webpack_app.liquid` is generated. The
format of these `webpack_<entry>.liquid` files is also changed. You now
use them e.g. as follows:

```
{% include 'webpack_app' %}

{% for file in app_webpack_css, cache: app_webpack_build_timestamp %}
{{ file | stylesheet_tag, media: "screen,print" }}
{% endfor %}

{% for file in app_webpack_js, cache: app_webpack_build_timestamp %}
{{ file | javascript_tag }}
{% endfor %}
```

__NOTE:__
If you need custom entries and you need this snippet-per-entry approach
to work, you need to specify your entries in `nimbu.js` as
`WEBPACK_ENTRY` instead of overriding the entries in `webpack.js`.